backend-python/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pymongo==4.10.1
firebase-admin==6.6.0
python-dotenv==1.0.1
